move button code, helpers into own files
