+++ src/components/shop/shop.js
@@ -1,5 +1,5 @@
 /**
- * Home is the main title screen for the game.
+ * Placeholder for shop screen.
  *
  * @module components/shop
  * @copyright BBC 2020
@@ -7,7 +7,6 @@
  * @license Apache-2.0
  */
 import { Screen } from "../../core/screen.js";
-
 import { scrollableList } from "../../core/layout/scrollable-list/scrollable-list.js";
 
 export class Shop extends Screen {
