+++ src/core/layout/scrollable-list/scrollable-list-helpers.js
@@ -1,0 +1,7 @@
+export const onClick = gelButton => {
+    console.log(`Clicked ${gelButton.config.id}`);
+};
+
+export const assetKey = (key, assetKeys) => {
+    return [assetKeys.prefix, key].join(".");
+};
