+++ src/core/layout/scrollable-list/scrollable-list.js
@@ -4,10 +4,11 @@
  * @author BBC Children's D+E
  * @license Apache-2.0 Apache-2.0
  */
-import { eventBus } from "../../event-bus.js"
+import { assetKey } from "./scrollable-list-helpers.js";
+import { createGelButton } from "./scrollable-list-buttons.js"
 
-export const scrollableList = (scene, config) => {
-    const panelConfig = getPanelConfig(scene, config);
+export const scrollableList = (scene) => {
+    const panelConfig = getPanelConfig(scene);
     const panel = scene.rexUI.add.scrollablePanel(panelConfig).layout();
     scene.input.topOnly = false;
     return panel;
@@ -94,57 +95,3 @@ export const createItem = (scene, item, config) => {
     });
     return label;
 };
-
-export const createGelButton = (scene, item, config) => {
-
-    const id = `shop_id_${item.id}`;
-
-    const gelConfig = {
-        gameButton: true,
-        accessibilityEnabled: true,
-        ariaLabel: item.ariaLabel, 
-        channel: config.eventChannel,
-        group: "middleCenter",
-        id,
-        key: config.assetKeys.itemBackground,
-        scene: config.assetKeys.prefix,
-    };
-
-    const gelButton = scene.add.gelButton(0, 0, gelConfig);
-
-    const callback = () => onClick(gelButton);
-    eventBus.subscribe({ callback, channel: gelConfig.channel, name: id });
-    
-    scaleButton(scene, gelButton, config);
-    return addOverlays(scene, gelButton, item, config);
-};
-
-export const addOverlays = (scene, gelButton, item, config) => {
-    const { offset, assetKeys: keys, font } = config;
-    const edge = gelButton.width / 2;
-    gelButton.overlays.set("background", scene.add.image(0, 0, assetKey(keys.itemBackground, keys)));
-    gelButton.overlays.set("icon", scene.add.image(-edge + offset.itemIconX, 0, item.icon));
-    gelButton.overlays.set("currencyIcon", scene.add.image(edge - offset.currencyIconX, 0, assetKey(keys.currency, keys)));
-
-    const fontStyle = { fontFamily: font.fontFamily, resolution: font.resolution };
-    gelButton.overlays.set("currencyAmount", scene.add.text(edge - offset.currencyTextX, -10, item.price, fontStyle));
-    const nameY = item.description ? offset.textY * 2 : offset.textY;
-    gelButton.overlays.set("itemName", scene.add.text(-edge + offset.textX, -nameY, item.name, { ...fontStyle, fontSize: 20 }));
-    gelButton.overlays.set("itemDescription", scene.add.text(-edge + offset.textX, 0, item.description ,fontStyle));
-
-    return gelButton;
-}
-
-export const scaleButton = (scene, button, config) => {
-    const safeArea = scene.layout.getSafeArea();
-    const scaleFactor = (safeArea.width - config.space * 2) / button.width;
-    button.setScale(scaleFactor);
-};
-
-const onClick = gelButton => {
-    console.log(`Clicked ${gelButton.config.id}`);
-}
-
-const assetKey = (key, assetKeys) => {
-    return [assetKeys.prefix, key].join(".");
-}
